Cleanup + Fixed autoupdating
- Moved csv related methos to `csv_helper.go`
- Moved other helper methods to `helpers.go`
- Moved auto update related methods to `auto_update.go`
- Added new `newVersionAvailable` property to App struct
- Add `UpdateAvailable` method so we can display info in the front-end
- Added `runMeElevated` and `restartSelf` methods
