+++ go.mod
@@ -9,7 +9,10 @@ require (
 	github.com/wailsapp/wails/v2 v2.7.1
 )
 
-require aead.dev/minisign v0.2.0 // indirect
+require (
+	aead.dev/minisign v0.2.0 // indirect
+	github.com/TheTitanrain/w32 v0.0.0-20180517000239-4f5cfb03fabf // indirect
+)
 
 require (
 	github.com/bep/debounce v1.2.1 // indirect
@@ -31,6 +34,7 @@ require (
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/rivo/uniseg v0.4.4 // indirect
 	github.com/samber/lo v1.38.1 // indirect
+	github.com/sqweek/dialog v0.0.0-20220809060634-e981b270ebbf
 	github.com/tkrajina/go-reflector v0.5.6 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/valyala/fasttemplate v1.2.2 // indirect
@@ -39,7 +43,7 @@ require (
 	golang.org/x/crypto v0.14.0 // indirect
 	golang.org/x/exp v0.0.0-20230522175609-2e198f4a06a1 // indirect
 	golang.org/x/net v0.17.0 // indirect
-	golang.org/x/sys v0.13.0 // indirect
+	golang.org/x/sys v0.13.0
 	golang.org/x/text v0.13.0 // indirect
 )
